feat: power menu and clipboard UI & UX improvements

+++ .local/bin/powermenu.sh
@@ -1,24 +1,30 @@
 #!/usr/bin/env bash
 
-shutdown="⏻  Shutdown"
-reboot="󰜉  Reboot"
-suspend="󰤄  Suspend"
-logout="󰍃  Logout"
-hyprland_reload="󰑓  Reload hyprland"
+shutdown="󰐥"
+reboot="󰜉"
+suspend="󰤄"
+logout="󰍃"
+hyprland_reload="󰑓"
 
 options="$shutdown\n$reboot\n$suspend\n$logout\n$hyprland_reload"
 
-chosen="$(echo -e "$options" | rofi -dmenu -i -p "Power Menu")"
+chosen="$(echo -e "$options" | rofi -dmenu -theme "$HOME/.config/rofi/powermenu.rasi")"
+
+[[ -z "$chosen" ]] && exit
+
+confirm_exit() {
+    echo -e "Confirm\nCancel" | rofi -dmenu -theme "$HOME/.config/rofi/confirm.rasi"
+}
 
 case $chosen in
     $shutdown)
-        systemctl poweroff ;;
+        if [[ $(confirm_exit) == "Confirm" ]]; then systemctl poweroff; fi ;;
     $reboot)
-        systemctl reboot ;;
+        if [[ $(confirm_exit) == "Confirm" ]]; then systemctl reboot; fi ;;
     $suspend)
         systemctl suspend ;;
     $logout)
-        hyprctl dispatch exit ;;
+         if [[ $(confirm_exit) == "Confirm" ]]; then hyprctl dispatch exit; fi ;;
     $hyprland_reload)
-        hyprctl reload ;;
+         hyprctl reload ;;
 esac
